refac: move ClassMethod manipulation into new class
